Implement input validation and display to user for both desktop and web client login functionality.
These changes are ment to enhance the login functionality by checking for errors in the users input such as null, blank, or empty inputs or invalid credentials and displays these errors to the user. This is done through catches and checks for response codes from ASP to the different systems that allow the UI to know what happened. Displays use UWP tooltips in desktop and spans with special styling in web client.