Implement adding and removing tags in desktop client
Implements adding and removing tags from a note in the desktop client using db and server changes from Aaron. Implemented an algorithm to manage tags for a note by seeing if the notes tag exists in the tags table and if not, it gets added, and then it gets the id of the tag and adds that to the note_tags table if and only if it wasn't already there.